# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.DetailedExamResultDTO:
    properties:
      correct_answer:
        example: A
        type: string
      domain_id:
        example: domain1
        type: string
      explanation:
        example: Digital signatures ensure non-repudiation.
        type: string
      is_correct:
        example: true
        type: boolean
      options: {}
      question_id:
        example: d1-q01
        type: string
      question_text:
        example: Which security control provides non-repudiation?
        type: string
      type:
        example: multiple_choice
        type: string
      user_answer:
        example: A
        type: string
    type: object
  dto.DomainScoreDTO:
    properties:
      correct_count:
        example: 3
        type: integer
      domain_id:
        example: domain1
        type: string
      percentage:
        example: 75
        type: number
      total_count:
        example: 4
        type: integer
    type: object
...
  dto.ErrorResponse:
    properties:
      code:
        example: BAD_REQUEST
        type: string
      error:
        example: Invalid request payload
        type: string
      message:
        example: Detailed error description
        type: string
    type: object
  dto.ExamSessionResponse:
    properties:
      questions:
        items:
          $ref: '#/definitions/dto.QuestionDTO'
        type: array
      session: {}
    type: object
...
  dto.QuestionDTO:
    properties:
      domain_id:
        example: domain1
        type: string
      options: {}
      order_index:
        example: 1
        type: integer
      question_id:
        example: d1-q01
        type: string
      question_text:
        example: Which security control provides non-repudiation?
        type: string
      type:
        example: multiple_choice
        type: string
      user_answer:
        example: A
        type: string
    type: object
...
  dto.SubmitAnswerRequest:
    properties:
      question_id:
        example: d1-q01
        type: string
      session_id:
        example: 987e6543-e21b-12d3-a456-426614174000
        type: string
      time_spent_seconds:
        example: 45
        type: integer
      user_answer:
        example: A
        type: string
    required:
    - question_id
    - session_id
    - user_answer
    type: object
  dto.SubmitAnswerResponse:
    properties:
      completed:
        example: true
        type: boolean
      correct_count:
        example: 25
        type: integer
      domain_scores:
        items:
          $ref: '#/definitions/dto.DomainScoreDTO'
        type: array
      message:
        example: Exam completed successfully!
        type: string
      results:
        items:
          $ref: '#/definitions/dto.DetailedExamResultDTO'
        type: array
      score:
        example: 83.33
        type: number
    type: object
...
host: localhost:8080
info:
  contact:
    name: SOC Trainer Development Team
  description: Automated Data-Driven Cyberattack Simulation for Blue Team Training
    platform REST API.
  license:
    name: MIT
  termsOfService: https://github.com/balliolon2/Automated_Data-Driven_Cyberattack_Simulation_for_Blue_Team_Training
  title: SOC Trainer REST API
  version: "1.0"
paths:
...
  /exams/post-test:
    post:
      consumes:
      - application/json
      description: Initializes a 30-question Post-Test with zero overlap from the
        Pre-Test questions.
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ExamSessionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Start post-test exam session
      tags:
      - Exams
  /exams/pre-test:
    post:
      consumes:
      - application/json
      description: Initializes a 30-question Pre-Test with required domain distribution
        (4/7/5/8/6). If an in-progress pre-test already exists, returns the existing
        session.
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ExamSessionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Start or resume pre-test exam session
      tags:
      - Exams
  /exams/sample:
    get:
      consumes:
      - application/json
      description: Retrieves a random question from the database question pool for
        preview.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.QuestionDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get a sample exam question
      tags:
      - Exams
  /exams/session:
    get:
      consumes:
      - application/json
      description: Recovers the active exam session and questions for the authenticated
        learner.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ExamSessionResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get active exam session
      tags:
      - Exams
  /exams/submit-answer:
    post:
      consumes:
      - application/json
      description: Commits an answer for an exam question and returns progression
        or final score results.
      parameters:
      - description: Exam Answer Submission
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SubmitAnswerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubmitAnswerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Submit an exam answer
      tags:
      - Exams
...
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      eyJhbGciOi..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"